requirements.txt
    Django==3.1
    djangorestframework==3.11.1
    flake8==3.8.3
    psycopg2==2.8.5
    psycopg2-binary==2.8.5
    psycopg2-pool==1.1
    Pillow==7.2.0
